requirements.txt
altair==4.2.2
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audio-recorder-streamlit==0.0.10
av==14.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.6.0
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.2
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
gitdb==4.0.12
GitPython==3.1.44
groq==0.24.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.0.2
onnxruntime==1.19.2
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.30.2
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.2
sniffio==1.3.1
streamlit==1.45.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.20.0
scipy>=1.7.0
librosa>=0.9.0
soundfile>=0.10.3
pyworld>=0.3.0
resemblyzer>=0.1.0
TTS>=0.22.0
